move business data to object
